1. add SemanticIndexBase 2. move common fucntions to data.py
