+++ examples/semantic_indexing/ance/run_ann_data_gen.py
@@ -17,6 +17,7 @@ from paddlenlp.utils.log import logger
 sys.path.append("../")
 from data import get_latest_checkpoint, get_latest_ann_data
 from data import convert_example, create_dataloader
+from data import gen_id2corpus, gen_text_file
 
 # yapf: disable
 parser = argparse.ArgumentParser()
@@ -41,31 +42,6 @@ parser.add_argument("--hnsw_max_elements", default=1000000, type=int, help="Reca
 
 args = parser.parse_args()
 # yapf: enable
-
-
-def get_semantic_embedding(data_loader, model):
-    """
-    Predicts the data labels.
-
-    Args:
-        model (obj:`paddle.nn.Layer`): A model to classify texts.
-        data_loader (obj:`List(Example)`): The processed data whose each element is a Example (numedtuple) object.
-            A Example object contains `text`(word_ids) and `se_len`(sequence length).
-
-    Returns:
-        results(obj:`List`): text embeddings generated by model.
-    """
-    model.eval()
-    with paddle.no_grad():
-        for batch_data in data_loader:
-            input_ids, token_type_ids = batch_data
-            input_ids = paddle.to_tensor(input_ids)
-            token_type_ids = paddle.to_tensor(token_type_ids)
-
-            text_embeddings = model.get_pooled_embedding(
-                input_ids, token_type_ids=token_type_ids)
-
-            yield text_embeddings.numpy()
 
 
 def build_index(args, data_loader, model):
@@ -98,7 +74,7 @@ def build_index(args, data_loader, model):
 
     all_embeddings = []
 
-    for text_embeddings in get_semantic_embedding(data_loader, model):
+    for text_embeddings in model.get_semantic_embedding(data_loader):
         all_embeddings.append(text_embeddings)
 
     all_embeddings = np.concatenate(all_embeddings, axis=0)
@@ -129,8 +105,8 @@ def generate_new_ann(args, data_loader_dict, checkpoint_path, latest_step_num):
                               model)
 
     logger.info("***** inference of query *****")
-    query_embedding = get_semantic_embedding(
-        data_loader_dict["text_data_loader"], model)
+    query_embedding = model.get_semantic_embedding(data_loader_dict[
+        "text_data_loader"])
 
     text_list = data_loader_dict["text_list"]
     id2corpus = data_loader_dict["id2corpus"]
@@ -166,49 +142,6 @@ def generate_new_ann(args, data_loader_dict, checkpoint_path, latest_step_num):
     succeed_flag_file = os.path.join(new_ann_data_path, "succeed_flag_file")
     open(succeed_flag_file, 'a').close()
     logger.info("finish generate ann data step:{}".format(latest_step_num))
-
-
-def get_semantic_embedding(data_loader, model):
-    """
-    Predicts the data labels.
-    Args:
-        model (obj:`paddle.nn.Layer`): A model to classify texts.
-        data_loader (obj:`List(Example)`): The processed data whose each element is a Example (numedtuple) object.
-            A Example object contains `text`(word_ids) and `se_len`(sequence length).
-
-    Returns:
-        results(obj:`List`): text embeddings generated by model.
-    """
-    model.eval()
-    with paddle.no_grad():
-        for batch_data in data_loader:
-            input_ids, token_type_ids = batch_data
-            input_ids = paddle.to_tensor(input_ids)
-            token_type_ids = paddle.to_tensor(token_type_ids)
-
-            text_embeddings = model.get_pooled_embedding(
-                input_ids, token_type_ids=token_type_ids)
-
-            yield text_embeddings.numpy()
-
-
-def gen_text_file(similar_text_pair_file):
-    text2similar_text = {}
-    texts = []
-    with open(similar_text_pair_file) as f:
-        for line in f:
-            text, similar_text = line.rstrip().split("\t")
-            text2similar_text[text] = similar_text
-            texts.append({"text": text})
-    return texts, text2similar_text
-
-
-def gen_id2corpus(corpus_file):
-    id2corpus = {}
-    with open(corpus_file) as f:
-        for idx, line in enumerate(f):
-            id2corpus[idx] = line.rstrip()
-    return id2corpus
 
 
 def build_data_loader(args, tokenizer):
