+++ examples/semantic_indexing/hardest_negative/model.py
@@ -12,53 +12,22 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import sys
+
 import numpy as np
 
 import paddle
 import paddle.nn as nn
 import paddle.nn.functional as F
 
+sys.path.append("../")
+from base_model import SemanticIndexBase
 
-class SemanticIndexHardestNeg(nn.Layer):
+
+class SemanticIndexHardestNeg(SemanticIndexBase):
     def __init__(self, pretrained_model, dropout=None, margin=0.3):
-        super().__init__()
-        self.ptm = pretrained_model
-        self.dropout = nn.Dropout(dropout if dropout is not None else 0.1)
+        super().__init__(pretrained_model, dropout)
         self.margin = margin
-
-    def get_pooled_embedding(self,
-                             input_ids,
-                             token_type_ids=None,
-                             position_ids=None,
-                             attention_mask=None):
-        _, cls_embedding = self.ptm(input_ids, token_type_ids, position_ids,
-                                    attention_mask)
-        cls_embedding = self.dropout(cls_embedding)
-        cls_embedding = F.normalize(cls_embedding, p=2, axis=-1)
-
-        return cls_embedding
-
-    def cosine_sim(self,
-                   query_input_ids,
-                   title_input_ids,
-                   query_token_type_ids=None,
-                   query_position_ids=None,
-                   query_attention_mask=None,
-                   title_token_type_ids=None,
-                   title_position_ids=None,
-                   title_attention_mask=None):
-
-        query_cls_embedding = self.get_pooled_embedding(
-            query_input_ids, query_token_type_ids, query_position_ids,
-            query_attention_mask)
-
-        title_cls_embedding = self.get_pooled_embedding(
-            title_input_ids, title_token_type_ids, title_position_ids,
-            title_attention_mask)
-
-        cosine_sim = paddle.sum(query_cls_embedding * title_cls_embedding,
-                                axis=-1).numpy()
-        return cosine_sim
 
     def forward(self,
                 query_input_ids,
