+++ examples/semantic_indexing/hardest_negative/run_ann.py
@@ -17,6 +17,7 @@
 from functools import partial
 import argparse
 import os
+import sys
 import random
 import time
 
@@ -29,7 +30,9 @@ from paddlenlp.data import Stack, Tuple, Pad
 from paddlenlp.datasets import load_dataset, MapDataset, load_dataset
 from paddlenlp.utils.log import logger
 
+sys.path.append("../")
 from data import convert_example, create_dataloader
+from data import gen_id2corpus, gen_text_file
 from model import SemanticIndexHardestNeg
 
 # yapf: disable
@@ -50,31 +53,6 @@ parser.add_argument("--hnsw_max_elements", default=1000000, type=int, help="Reca
 parser.add_argument('--device', choices=['cpu', 'gpu'], default="gpu", help="Select which device to train model, defaults to gpu.")
 args = parser.parse_args()
 # yapf: enable
-
-
-def get_semantic_embedding(data_loader, model):
-    """
-    Predicts the data labels.
-
-    Args:
-        model (obj:`paddle.nn.Layer`): A model to classify texts.
-        data_loader (obj:`List(Example)`): The processed data whose each element is a Example (numedtuple) object.
-            A Example object contains `text`(word_ids) and `se_len`(sequence length).
-
-    Returns:
-        results(obj:`List`): text embeddings generated by model.
-    """
-    model.eval()
-    with paddle.no_grad():
-        for batch_data in data_loader:
-            input_ids, token_type_ids = batch_data
-            input_ids = paddle.to_tensor(input_ids)
-            token_type_ids = paddle.to_tensor(token_type_ids)
-
-            text_embeddings = model.get_pooled_embedding(
-                input_ids, token_type_ids=token_type_ids)
-
-            yield text_embeddings.numpy()
 
 
 def build_index(data_loader, model):
@@ -107,7 +85,7 @@ def build_index(data_loader, model):
 
     all_embeddings = []
 
-    for text_embeddings in get_semantic_embedding(data_loader, model):
+    for text_embeddings in model.get_semantic_embedding(data_loader):
         all_embeddings.append(text_embeddings)
 
     all_embeddings = np.concatenate(all_embeddings, axis=0)
@@ -116,25 +94,6 @@ def build_index(data_loader, model):
     logger.info("Total index number:{}".format(index.get_current_count()))
 
     return index
-
-
-def id2corpus(corpus_file):
-    id2corpus = {}
-    with open(corpus_file) as f:
-        for idx, line in enumerate(f):
-            id2corpus[idx] = line.rstrip()
-    return id2corpus
-
-
-def gen_text_file(similar_text_pair_file):
-    text2similar_text = {}
-    texts = []
-    with open(args.similar_text_pair_file) as f:
-        for line in f:
-            text, similar_text = line.rstrip().split("\t")
-            text2similar_text[text] = similar_text
-            texts.append({"text": text})
-    return texts, text2similar_text
 
 
 if __name__ == "__main__":
@@ -163,7 +122,7 @@ if __name__ == "__main__":
         model.set_dict(state_dict)
         logger.info("Loaded parameters from %s" % args.params_path)
 
-    id2corpus = id2corpus(args.corpus_file)
+    id2corpus = gen_id2corpus(args.corpus_file)
 
     # conver_example function's input must be dict
     corpus_list = [{idx: text} for idx, text in id2corpus.items()]
@@ -189,7 +148,7 @@ if __name__ == "__main__":
         batchify_fn=batchify_fn,
         trans_fn=trans_func)
 
-    query_embedding = get_semantic_embedding(query_data_loader, model)
+    query_embedding = model.get_semantic_embedding(query_data_loader)
 
     recall_result_dir = "recall_result"
     if not os.path.exists(recall_result_dir):
